Refactored code

+++ src/ball.rs
@@ -5,21 +5,21 @@ use crate::{BALL_COLOR, BALL_SIZE};
 use piston::UpdateArgs;
 
 pub struct Ball {
-    pub x: f32,
-    pub x_velocity: f32,
-    pub y: f32,
-    pub y_velocity: f32,
+    pub x: f64,
+    pub x_velocity: f64,
+    pub y: f64,
+    pub y_velocity: f64,
 }
 
 impl Ball {
     /// Moves the ball with the velocity and bounces the ball of the "walls" specified using the y_range if they hit a wall.
-    pub fn update(&mut self, update_args: &UpdateArgs, mut y_range: Range<f32>) {
+    pub fn update(&mut self, update_args: &UpdateArgs, mut y_range: Range<f64>) {
         // Move the ball the amount of velocity multiplied by the delta time.
-        self.x += self.x_velocity * (update_args.dt as f32);
-        self.y += self.y_velocity * (update_args.dt as f32);
+        self.x += self.x_velocity * (update_args.dt as f64);
+        self.y += self.y_velocity * (update_args.dt as f64);
 
         // Subtract the ball height from the range so that the ball won't go off the screen
-        y_range.end -= BALL_SIZE.1 as f32;
+        y_range.end -= BALL_SIZE.1 as f64;
 
         if self.y < y_range.start {
             // If the y position is smaller
@@ -39,8 +39,8 @@ impl Ball {
     /// Renders the ball.
     pub fn render<G>(&self, context: &Context, graphics: &mut G) where G: Graphics {
         let rect = [
-            self.x as f64,
-            self.y as f64,
+            self.x,
+            self.y,
             BALL_SIZE.0 as f64,
             BALL_SIZE.1 as f64,
         ];
